backend/requirements.txt
flask
flask-cors
cohere
pandas
openpyxl
python-dotenv
